pyproject.toml
[tool.poetry]
name = "crypto-analysis"
version = "0.1.0"
description = ""
authors = ["Thomás Freire <thomas.fsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandera = "^0.18.3"
duckdb = "^0.10.0"
mkdocs = "^1.5.3"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocstrings = {extras = ["python"], version = "^0.24.1"}
taskipy = "^1.12.2"
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
python-coinmarketcap = "^0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
